design: common base for worker and hub
We should have a common base with worker functions to make it
cleaner for a hub to serve as a worker.
